server/requirements.txt
protobuf==5.28.3
python-dotenv==1.0.1
google-generativeai
flask
flask_cors